feat: add user/admin roles

+++ app/services/task.py
@@ -1,4 +1,6 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
+from app.core.roles import Role
 from app.exception.task import TaskNotFound
 from app.models.user import UserORM
 from app.repositories.task import TaskRepository
@@ -22,11 +24,21 @@ class TaskService:
         self.db.commit()
         return TaskSchema.model_validate(create_tasks_orm)
     
-    def update_task(self, task_id: str, task_update: TaskUpdateSchema) -> TaskSchema:
+    def update_task(
+        self, task_id: str, 
+        task_update: TaskUpdateSchema,
+        user: UserORM
+    ) -> TaskSchema:
         task_for_update = self.task_repository.get_by_id(task_id=task_id)
         
         if not task_for_update:
             raise TaskNotFound(f"The task with id {task_id} not found.")
+        
+        if task_for_update.user_id != user.id and user.role != Role.admin:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Forbidden"
+            )
         
         if task_update.title is not None:
             task_for_update.title = task_update.title
@@ -37,9 +49,16 @@ class TaskService:
         
         return TaskSchema.model_validate(task_for_update)
     
-    def delete_task(self, task_id: str) -> None:
+    def delete_task(self, task_id: str, user: UserORM) -> None:
         task_for_delete = self.task_repository.get_by_id(task_id=task_id)
         if not task_for_delete:
             raise TaskNotFound(f"The task with id {task_id} not found.")
+        
+        if task_for_delete.user_id != user.id and user.role != Role.admin:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Forbidden"
+            )
+            
         self.task_repository.delete_task(task_for_delete)
         self.db.commit()
